Remove ContentDialogOverlay
