+++ Ryujinx.Ava/UI/Helpers/ContentDialogHelper.cs
@@ -1,5 +1,6 @@
 using Avalonia;
 using Avalonia.Controls;
+using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Media;
 using Avalonia.Threading;
 using FluentAvalonia.Core;
@@ -62,7 +63,7 @@ namespace Ryujinx.Ava.UI.Helpers
                 contentDialog.PrimaryButtonClick += deferCloseAction;
             }
 
-            await ContentDialogOverlay.ShowAsync(contentDialog);
+            await ShowAsync(contentDialog);
 
             return result;
         }
@@ -338,6 +339,99 @@ namespace Ryujinx.Ava.UI.Helpers
             }
 
             return string.Empty;
+        }
+
+        public static async Task<ContentDialogResult> ShowAsync(ContentDialog contentDialog)
+        {
+            ContentDialogResult result;
+
+            ContentDialogOverlayWindow contentDialogOverlayWindow = null;
+
+            Window parent = GetMainWindow();
+
+            if (parent.IsActive && parent is MainWindow window && window.ViewModel.IsGameRunning)
+            {
+                contentDialogOverlayWindow = new()
+                {
+                    Height        = parent.Bounds.Height,
+                    Width         = parent.Bounds.Width,
+                    Position      = parent.PointToScreen(new Point()),
+                    ShowInTaskbar = false
+                };
+
+                parent.PositionChanged += OverlayOnPositionChanged;
+
+                void OverlayOnPositionChanged(object sender, PixelPointEventArgs e)
+                {
+                    contentDialogOverlayWindow.Position = parent.PointToScreen(new Point());
+                }
+
+                contentDialogOverlayWindow.ContentDialog = contentDialog;
+
+                bool opened = false;
+
+                contentDialogOverlayWindow.Opened += OverlayOnActivated;
+
+                async void OverlayOnActivated(object sender, EventArgs e)
+                {
+                    if (opened)
+                    {
+                        return;
+                    }
+
+                    opened = true;
+
+                    contentDialogOverlayWindow.Position = parent.PointToScreen(new Point());
+
+                    result = await ShowDialog();
+                }
+
+                result = await contentDialogOverlayWindow.ShowDialog<ContentDialogResult>(parent);
+            }
+            else
+            {
+                result = await ShowDialog();
+            }
+
+            async Task<ContentDialogResult> ShowDialog()
+            {
+                if (contentDialogOverlayWindow is not null)
+                {
+                    result = await contentDialog.ShowAsync(contentDialogOverlayWindow);
+
+                    contentDialogOverlayWindow!.Close();
+                }
+                else
+                {
+                    result = await contentDialog.ShowAsync();
+                }
+
+                return result;
+            }
+
+            if (contentDialogOverlayWindow is not null)
+            {
+                contentDialogOverlayWindow.Content = null;
+                contentDialogOverlayWindow.Close();
+            }
+
+            return result;
+        }
+
+        private static Window GetMainWindow()
+        {
+            if (Application.Current.ApplicationLifetime is IClassicDesktopStyleApplicationLifetime al)
+            {
+                foreach (Window item in al.Windows)
+                {
+                    if (item.IsActive && item is MainWindow window)
+                    {
+                        return window;
+                    }
+                }
+            }
+
+            return null;
         }
     }
 }
